-Created PillTrackerAPI.kt Class
